+++ src/components/User/Users/components/UserDetails/UserDetails.Kyc.vue
@@ -5,150 +5,124 @@
     </h3>
     <div class="user-details-account__info">
       <p>Can be uploaded one of the documents:</p>
-      <p> - Valid passport;</p>
-      <p>- Driving license - full, not provisional, with a color photograph;</p>
+      <p> - Valid Passport;</p>
+      <p> - Valid Identity card;</p>
+      <p> - Valid Residence permit;</p>
       <p>
-        - Valid National ID card;
-      </p><p />
-      <p>
-        - In all cases handwritten documents or those in a foreign script
-        that cannot be read, i.e Arabic or Cyrillic
-        will not be accepted unless translated and notarized.
+        - Driving license - full, not provisional, with a color photograph;
       </p>
     </div>
-
-    <template v-if="isLoaded">
-      <div class="user-details-account__docs">
-        <div class="user-details-account__doc-view-wrp">
-          <h3>{{ ID_DOCUMENTS_VERBOSE[kyc.idDocumentType] }}</h3>
-          <user-doc-getter
-            class="user-details-account__doc-view"
-            :user-id="user.id"
-            :file-key="kyc.documents.kycIdDocument"
-          />
-        </div>
-        <div
-          v-if="config.FEATURES.PHOTO_VERIFICATION"
-          class="user-details-account__doc-view-wrp"
-        >
-          <h3>Photo with code: {{ user.address.slice(1, 6) }}</h3>
-          <user-doc-getter
-            class="user-details-account__doc-view"
-            :user-id="user.address"
-            :file-key="kyc.documents.bravo || kyc.documents.kycSelfie"
-          />
-        </div>
-        <div
-          v-if="kyc.documents.kycAvatar"
-          class="user-details-account__doc-view-wrp"
-        >
-          <h3>Avatar</h3>
-          <user-doc-getter
-            class="user-details-account__doc-view"
-            :file-key="kyc.documents.kycAvatar"
-          />
-        </div>
-      </div>
-
-      <ul class="key-value-list">
-        <li>
-          <span>First name</span>
-          <span :title="kyc.firstName">
-            {{ kyc.firstName }}
-          </span>
-        </li>
-        <li>
-          <span>Last name</span>
-          <span :title="kyc.lastName">
-            {{ kyc.lastName }}
-          </span>
-        </li>
-        <li>
-          <span>ID Document type</span>
-          <span :title="ID_DOCUMENTS_VERBOSE[kyc.idDocumentType]">
-            {{ ID_DOCUMENTS_VERBOSE[kyc.idDocumentType] }}
-          </span>
-        </li>
-        <li>
-          <span>ID Document</span>
-          <span>
-            <user-doc-link-getter
-              :file-key="kyc.documents.kycIdDocument"
-            >
-              Open file
-            </user-doc-link-getter>
-          </span>
-        </li>
-        <li v-if="kyc.documents.kycAvatar">
-          <span>Avatar</span>
-          <span>
-            <user-doc-link-getter :file-key="kyc.documents.kycAvatar">
-              Open file
-            </user-doc-link-getter>
-          </span>
-        </li>
-        <li v-if="kyc.documents.bravo">
-          <span>Photo with verification code</span>
-          <span>
-            <user-doc-link-getter
-              :file-key="kyc.documents.bravo"
-            >
-              Open file
-            </user-doc-link-getter>
-          </span>
-        </li>
-        <li v-if="kyc.documents.kycSelfie">
-          <span>Photo with verification code</span>
-          <span>
-            <user-doc-link-getter
-              :file-key="kyc.documents.kycSelfie"
-            >
-              Open file
-            </user-doc-link-getter>
-          </span>
-        </li>
-        <li
-          class="code-details"
-          v-if="kyc.documents.bravo || kyc.documents.kycSelfie"
-        >
-          Code: {{ user.address.slice(1, 6) }}
-        </li>
-      </ul>
-      <div class="user-details-account__info">
-        <p>Can be uploaded one of the documents:</p>
-        <p> - Driving license if not already uploaded on the previous step;</p>
-        <p>- Recent utility bill (not older than 6 months old);</p>
-        <p>
-          - Bank statement or credit card statement (not older than 6 months);
-        </p><p />
-        <p>
-          - Bank reference letter (not older than 6 months);
-        </p><p />
-        <p>- Signed Tenancy/Rental Agreement.</p>
-      </div>
-    </template>
-
-    <template v-else-if="isFailed">
-      <p class="danger">
-        An error occurred. Please try again later.
-      </p>
-    </template>
-
-    <template v-else>
-      <p>Loading...</p>
-    </template>
+    <ul class="user-details-account__key-value-list key-value-list">
+      <li>
+        <span>First name</span>
+        <span :title="kyc.firstName">
+          {{ kyc.firstName }}
+        </span>
+      </li>
+      <li>
+        <span>Last name</span>
+        <span :title="kyc.lastName">
+          {{ kyc.lastName }}
+        </span>
+      </li>
+      <li>
+        <span>Date of birth</span>
+        <span :title="kyc.dateOfBirth">
+          {{ kyc.dateOfBirth }}
+        </span>
+      </li>
+      <li>
+        <span>Address Line 1</span>
+        <span :title="kyc.address.line1">
+          {{ kyc.address.line1 }}
+        </span>
+      </li>
+      <li>
+        <span>Address Line 2</span>
+        <span :title="kyc.address.line2">
+          {{ kyc.address.line2 }}
+        </span>
+      </li>
+      <li>
+        <span>City</span>
+        <span :title="kyc.address.city">
+          {{ kyc.address.city }}
+        </span>
+      </li>
+      <li>
+        <span>Country</span>
+        <span :title="country">
+          {{ country }}
+        </span>
+      </li>
+      <li>
+        <span>State</span>
+        <span :title="kyc.address.state">
+          {{ kyc.address.state }}
+        </span>
+      </li>
+      <li>
+        <span>ID Document type</span>
+        <span :title="ID_DOCUMENTS_VERBOSE[kyc.idDocumentType]">
+          {{ ID_DOCUMENTS_VERBOSE[kyc.idDocumentType] }}
+        </span>
+      </li>
+      <li>
+        <span v-if="kyc.documents.kycIdDocument.back">
+          ID Document front side:
+        </span>
+        <span v-else>
+          ID Document
+        </span>
+        <span>
+          <user-doc-link-getter
+            :file-key="kyc.documents.kycIdDocument.face"
+          >
+            Open file
+          </user-doc-link-getter>
+        </span>
+      </li>
+      <li v-if="kyc.documents.kycIdDocument.back">
+        <span>ID Document back side: </span>
+        <span>
+          <user-doc-link-getter
+            :file-key="kyc.documents.kycIdDocument.back"
+          >
+            Open file
+          </user-doc-link-getter>
+        </span>
+      </li>
+      <li v-if="kyc.documents.kycAvatar">
+        <span>Avatar</span>
+        <span>
+          <user-doc-link-getter :file-key="kyc.documents.kycAvatar">
+            Open file
+          </user-doc-link-getter>
+        </span>
+      </li>
+      <li v-if="kyc.documents.kycSelfie">
+        <span>Photo with verification code</span>
+        <span>
+          <user-doc-link-getter
+            :file-key="kyc.documents.kycSelfie"
+          >
+            Open file
+          </user-doc-link-getter>
+        </span>
+      </li>
+      <li
+        class="code-details"
+        v-if="kyc.documents.kycSelfie"
+      >
+        Code: {{ user.address.slice(1, 6) }}
+      </li>
+    </ul>
   </div>
 </template>
 
 <script>
-import { UserDocGetter, UserDocLinkGetter } from '@comcom/getters'
-
-import { fromKycTemplate } from '../../../../../utils/kyc-tempater'
-import deepCamelCase from 'camelcase-keys-deep'
-
-import config from '@/config'
-import { Sdk } from '@/sdk'
-import { ErrorHandler } from '@/utils/ErrorHandler'
+import { UserDocLinkGetter } from '@comcom/getters'
+import { byAlpha2 } from 'iso-country-codes'
 
 const ID_DOCUMENTS_VERBOSE = {
   passport: 'Passport',
@@ -159,50 +133,32 @@ const ID_DOCUMENTS_VERBOSE = {
 
 export default {
   components: {
-    UserDocGetter,
     UserDocLinkGetter,
   },
 
   props: {
-    user: { type: Object, required: true },
-    blobId: { type: String, required: true },
+    kyc: {
+      type: Object,
+      required: true,
+    },
+    user: {
+      type: Object,
+      required: true,
+    },
   },
 
   data () {
     return {
-      kyc: {},
-      isLoaded: false,
-      isFailed: false,
-      config,
       ID_DOCUMENTS_VERBOSE,
     }
   },
 
-  watch: {
-    user () { this.getKyc() },
-    blobId () { this.getKyc() },
-  },
-
-  created () {
-    this.getKyc()
-  },
-
-  methods: {
-    async getKyc () {
-      this.isLoaded = false
-      this.isFailed = false
-
-      try {
-        const response = await Sdk.api.blobs.get(this.blobId, this.user.address)
-        const kycFormResponse = response.data
-        this.kyc = deepCamelCase(
-          fromKycTemplate(JSON.parse(kycFormResponse.value))
-        )
-        this.isLoaded = true
-      } catch (error) {
-        ErrorHandler.processWithoutFeedback(error)
-        this.isFailed = true
-      }
+  computed: {
+    country () {
+      const country = Object
+        .keys(byAlpha2)
+        .find(country => country === this.kyc.address.country)
+      return byAlpha2[country].name
     },
   },
 }
@@ -237,4 +193,10 @@ export default {
 .user-details-account__info {
   margin-bottom: 3rem;
 }
+
+.user-details-account__key-value-list-id-document {
+  display: flex;
+  flex-direction: column;
+}
+
 </style>