+++ src/components/User/Users/components/UserDetails/UserDetails.vue
@@ -8,39 +8,37 @@
           class="user-details__section"
           v-if="requestToReview.state"
         >
-          <h3>
-            Current request state
-          </h3>
-          <p
-            :class="`user-details__state-info
-                      user-details__state-info--${requestToReview.state}`">
-            {{ requestToReview.state }}
-          </p>
-
+          <ul class="key-value-list">
+            <li>
+              <h3>Current request state</h3>
+              <p
+                :class="`user-details__state-info
+                         user-details__state-info--${requestToReview.state}`">
+                {{ requestToReview.state }}
+              </p>
+            </li>
+            <li>
+              <h3>Role to set</h3>
+              <p class="user-details__role-info">
+                {{
+                  ROLE_TYPE_VERBOSE[requestToReview.accountRoleToSet
+                    || verifiedRequest.accountRoleToSet]
+                }}
+              </p>
+            </li>
+          </ul>
           <p v-if="requestToReview.isRejected">
             Reason: {{ requestToReview.rejectReason }}
           </p>
 
-          <p
-            class="user-details__heading"
-            v-if="requestToReview.isRejected"
-          >
-            <span>External rejection details</span>
-            <button
-              class="app__btn-secondary app__btn-secondary--iconed"
-              @click="isShownExternal = !isShownExternal"
-            >
-              <mdi-chevron-up-icon v-if="isShownExternal" />
-              <mdi-chevron-down-icon v-else />
-            </button>
-          </p>
-
-          <!-- eslint-disable vue/no-v-html -->
-          <p
-            v-if="isShownExternal"
-            v-html="externalDetails"
-          />
-          <!-- eslint-enable vue/no-v-html -->
+          <template v-if="requestToReview.externalDetails">
+            <div class="user-details__ext-details-wrp">
+              <h3>External details (provided by external services)</h3>
+              <external-details-viewer
+                :external-details="requestToReview.externalDetails"
+              />
+            </div>
+          </template>
         </section>
 
         <section class="user-details__section">
@@ -53,15 +51,39 @@
 
         <template v-if="requestToReview.state">
           <section class="user-details__section">
-            <kyc-general-section
-              v-if="requestToReview.accountRoleToSet === ACCOUNT_ROLES.general"
-              :user="user"
-              :blob-id="requestToReview.blobId"
-            />
-
-            <!-- eslint-disable max-len -->
+            <template v-if="isKycLoaded">
+              <general-kyc-viewer
+                v-if="
+                  requestToReview.accountRoleToSet === ACCOUNT_ROLES.general
+                "
+                :kyc="kyc"
+                :user="user" />
+              <verified-kyc-viewer
+                v-if="
+                  requestToReview.accountRoleToSet === ACCOUNT_ROLES.usVerified
+                "
+                :kyc="kyc"
+                :user="user" />
+              <accredited-kyc-viewer
+                v-if="
+                  requestToReview.accountRoleToSet ===
+                    ACCOUNT_ROLES.usAccredited
+                "
+                :kyc="kyc"
+                :user="user" />
+            </template>
+            <template v-else-if="isKycLoadFailed">
+              <p class="danger">
+                An error occurred. Please try again later.
+              </p>
+            </template>
+            <template v-else>
+              <p>Loading...</p>
+            </template>
             <kyc-syndicate-section
-              v-if="requestToReview.accountRoleToSet === ACCOUNT_ROLES.corporate"
+              v-if="
+                requestToReview.accountRoleToSet === ACCOUNT_ROLES.corporate
+              "
               :user="user"
               :blob-id="requestToReview.blobId"
             />
@@ -75,14 +97,34 @@
             v-if="verifiedRequest.accountRoleToSet !== ACCOUNT_ROLES.notVerified && !isUserBlocked"
             class="user-details__section"
           >
-            <h2>Previous approved KYC Request</h2>
-            <kyc-general-section
-              v-if="verifiedRequest.accountRoleToSet === ACCOUNT_ROLES.general"
-              :user="user"
-              :blob-id="verifiedRequest.blobId"
-            />
+            <template v-if="isKycLoaded">
+              <h2>Previous approved KYC Request</h2>
+              <general-kyc-viewer
+                v-if="verifiedRequest.accountRoleToSet === ACCOUNT_ROLES.general"
+                :kyc="kyc"
+                :user="user"
+              />
+              <verified-kyc-viewer
+                v-if="verifiedRequest.accountRoleToSet === ACCOUNT_ROLES.usVerified"
+                :kyc="kyc"
+                :user="user"
+              />
+              <accredited-kyc-viewer
+                v-if="verifiedRequest.accountRoleToSet === ACCOUNT_ROLES.usAccredited"
+                :kyc="kyc"
+                :user="user"
+              />
+            </template>
+            <template v-else-if="isKycLoadFailed">
+              <p class="danger">
+                An error occurred. Please try again later.
+              </p>
+            </template>
+            <template v-else>
+              <p>Loading...</p>
+            </template>
             <kyc-syndicate-section
-              v-else-if="verifiedRequest.accountRoleToSet === ACCOUNT_ROLES.corporate"
+              v-if="verifiedRequest.accountRoleToSet === ACCOUNT_ROLES.corporate"
               :user="user"
               :blob-id="verifiedRequest.blobId"
             />
@@ -152,21 +194,35 @@
 <script>
 import AccountSection from './UserDetails.Account'
 
-import KycGeneralSection from './UserDetails.Kyc'
 import KycSyndicateSection from '@/components/User/Sales/components/SaleManager/SaleManager.SyndicateTab'
+
+import GeneralKycViewer from './UserDetails.GeneralKycViewer'
+import VerifiedKycViewer from './UserDetails.VerifiedKycViewer'
+import AccreditedKycViewer from './UserDetails.AccreditedKycViewer'
 
 import RequestActions from './UserDetails.Request'
 import ResetActions from './UserDetails.Reset'
 import BlockActions from './UserDetails.Block'
 
-import { unCamelCase } from '@/utils/un-camel-case'
+import ExternalDetailsViewer from './UserDetails.ExternalDetailsViewer'
 
 import { ApiCallerFactory } from '@/api-caller-factory'
 import { ErrorHandler } from '@/utils/ErrorHandler'
 
 import { ChangeRoleRequest } from '@/api/responseHandlers/requests/ChangeRoleRequest'
+import { fromKycTemplate } from '../../../../../utils/kyc-tempater'
+import deepCamelCase from 'camelcase-keys-deep'
+import { Sdk } from '@/sdk'
 
 import config from '@/config'
+
+const ROLE_TYPE_VERBOSE = {
+  [ config.ACCOUNT_ROLES.general ]: 'General',
+  [ config.ACCOUNT_ROLES.usVerified ]: 'US Verified',
+  [ config.ACCOUNT_ROLES.usAccredited ]: 'US Accredited',
+  [ config.ACCOUNT_ROLES.corporate ]: 'Corporate',
+  [ config.ACCOUNT_ROLES.notVerified ]: 'Not Verified',
+}
 
 const OPERATION_TYPE = {
   createKycRequest: '22',
@@ -178,11 +234,14 @@ const EVENTS = {
 export default {
   components: {
     AccountSection,
-    KycGeneralSection,
+    ExternalDetailsViewer,
     KycSyndicateSection,
     RequestActions,
     ResetActions,
     BlockActions,
+    AccreditedKycViewer,
+    VerifiedKycViewer,
+    GeneralKycViewer,
   },
 
   props: {
@@ -196,39 +255,31 @@ export default {
       isLoaded: false,
       isFailed: false,
       isPending: false,
-      isShownExternal: false,
+      isKycLoaded: false,
+      isKycLoadFailed: false,
       user: {},
       requests: [],
-      verifiedRequest: new ChangeRoleRequest({}),
+      verifiedRequest: {},
+      kyc: {},
+      ROLE_TYPE_VERBOSE,
     }
   },
 
   computed: {
-    externalDetails () {
-      if (!this.requestToReview.externalDetails) return ''
-      return this.requestToReview.externalDetails
-        .map(detail =>
-          Object.entries(detail)
-            .map(([key, value]) => `${unCamelCase(key)}: ${value}`)
-            .join(' ')
-        )
-        .filter(value => value)
-        .join('<br>')
-    },
-
     isUserBlocked () {
       return this.user.role === config.ACCOUNT_ROLES.blocked
     },
 
     latestApprovedRequest () {
-      return this.requests.find(item => item.isApproved) ||
-        new ChangeRoleRequest({})
+      return (
+        this.requests.find(item => item.isApproved)
+      ) || new ChangeRoleRequest({})
     },
 
     requestToReview () {
-      return this.requests
-        .find(item => item.isPending || item.isRejected) ||
-        new ChangeRoleRequest({})
+      return (
+        this.requests.find(item => item.isPending || item.isRejected)
+      ) || new ChangeRoleRequest({})
     },
 
     latestBlockedRequest () {
@@ -254,6 +305,12 @@ export default {
 
   async created () {
     await this.getUser()
+
+    if (this.requestToReview.state) {
+      await this.getKyc(this.requestToReview.blobId)
+    } else if (this.verifiedRequest.state) {
+      await this.getKyc(this.verifiedRequest.blobId)
+    }
   },
 
   methods: {
@@ -313,6 +370,23 @@ export default {
         this.$emit(EVENTS.reviewed)
       }, 5000)
     },
+
+    async getKyc (blodId) {
+      this.isKycLoaded = false
+      this.isKycLoadFailed = false
+
+      try {
+        const response = await Sdk.api.blobs.get(blodId, this.user.address)
+        const kycFormResponse = response.data
+        this.kyc = deepCamelCase(
+          fromKycTemplate(JSON.parse(kycFormResponse.value))
+        )
+        this.isKycLoaded = true
+      } catch (error) {
+        ErrorHandler.process(error)
+        this.isKycLoadFailed = true
+      }
+    },
   },
 }
 </script>
@@ -362,5 +436,9 @@ export default {
   &--approved { color: $color-success; }
   &--pending { color: $color-active; }
   &--rejected { color: $color-danger; }
+}
+
+.user-details__ext-details-wrp {
+  margin: 2rem 0;
 }
 </style>